Drop outdated components
These components are either already archived or otherwise outdated, e.g.,
- rely on Java 7
- were already dropped from frequent or even aggregator builds.

Note that some of them should be even archived on GitHub as they
are no longer maintained.

Detected in the course of support-and-care/maven-support-and-care#77.
Cf. https://github.com/support-and-care/maven-support-and-care/blob/main/src/docs/epics/77-maven-due-diligence/index.adoc#archived-components